added a working search bar on the right view
