+++ app/src/main/res/layout/content_select_food.xml
@@ -11,18 +11,11 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Search for food"
-            android:id="@+id/foodSearch"
-            />
 
         <android.support.v7.widget.RecyclerView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:id="@+id/selectFoodRecyclerView"
-            android:layout_below="@id/foodSearch"
             >
 
         </android.support.v7.widget.RecyclerView>
